Refactor interrupts
Moved interrupt related things from init.h to interrupt.h
